+++ src/tree/FileTreeItem.ts
@@ -1,15 +1,21 @@
 import DeviceTreeItem from './DeviceTreeItem'
+import FolderTreeItem from './FolderTreeItem'
 import { TreeItem } from 'vscode'
 
 export default class FileTreeItem extends TreeItem {
-	public readonly parent: DeviceTreeItem
-	public readonly name: string
+	public readonly path: string
 
-	constructor(name: string, size: number, parent: DeviceTreeItem) {
-		super(name)
+	constructor(
+		public readonly name: string,
+		public readonly size: number,
+		public readonly parent: DeviceTreeItem | FolderTreeItem,
+	) {
+		const nameToShow = name.includes('/') ? name.split('/')[1] : name
+		super(nameToShow)
 
 		this.parent = parent
 		this.name = name
+		this.path = parent.path
 
 		this.id = `nodemcu-file-${parent.path}-${name}`
 		this.tooltip = `${size}`